IsVendor: move RE collation to code generation phase
